+++ packages/saas-ui-theme/stories/chakra-ui.stories.tsx
@@ -46,20 +46,14 @@ import {
   CardFooter,
   Image,
   useClipboard,
+  Divider,
 } from '@chakra-ui/react'
 import {
   Select,
   PropertyList,
   Property,
-  // List,
-  // ListItem,
-  // ListItemIcon,
   PersonaAvatar,
-  // ListItemLabel,
-  // ListItemTertiary,
-  // ListItemAction,
   SearchInput,
-  Divider,
   Link,
   useSnackbar,
 } from '@saas-ui/react'