+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,14 @@
 # saas-ui-storybook
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [df98007a]
+- Updated dependencies [9b609119]
+  - @saas-ui/react@2.0.2
+  - @saas-ui/storybook-addon@1.1.0
+
 ## 1.0.2
 
 ### Patch Changes